+++ tests/Unit/Support/PipelineTest.php
@@ -17,6 +17,7 @@ use OzanKurt\Tracker\Repositories\SessionRepository;
 use OzanKurt\Tracker\Support\BotFilter;
 use OzanKurt\Tracker\Support\Enricher;
 use OzanKurt\Tracker\Support\Pipeline;
+use OzanKurt\Tracker\Support\PrivacyFilter;
 use OzanKurt\Tracker\Support\RefererParser;
 
 beforeEach(function () {
@@ -49,7 +50,7 @@ function makePipelineForPipelineTest(): Pipeline
 
     return new Pipeline(
         botFilter: new BotFilter,
-        enricher: new Enricher($geo, new RefererParser),
+        enricher: new Enricher($geo, new RefererParser, new PrivacyFilter),
         repositories: $repos,
     );
 }